simplify year entity, but tests don't pass

+++ src/main/java/org/akcjamis/webapp/repository/EventRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 /**
  * Spring Data JPA repository for the Event entity.
  */
-public interface EventRepository extends JpaRepository<Event,Long> {
+public interface EventRepository extends JpaRepository<Event,Short> {
 
 }
